requirements.txt
Django==3.2.19
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
celery==5.2.7
ipython==8.14.0
python-dotenv==1.0.0
drf-yasg==1.21.5
mysqlclient==2.1.1
python-magic==0.4.27
blurhash-python==1.2.0
Faker==18.10.0
amqp==5.1.1
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
billiard==3.6.4.0
blurhash-python==1.2.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.2.19
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.5
executing==1.2.0
Faker==18.10.0
idna==3.4
inflection==0.5.1
ipython==8.14.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mysqlclient==2.1.1
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.0
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
redis==4.5.5
